Chore : 안 쓰는 파일 삭제
